fix(Webserver): fix thumbnail ratio calculation not using floating point division

+++ Hayden.WebServer/Controllers/Api/PostingApi.cs
@@ -368,12 +368,12 @@ namespace Hayden.WebServer.Controllers.Api
                         if (tempFile.ImageWidth > tempFile.ImageHeight)
                         {
                             width = 125;
-                            height = (int)((tempFile.ImageHeight / tempFile.ImageWidth) * 125);
+                            height = (int)(((double)tempFile.ImageHeight / tempFile.ImageWidth) * 125);
                         }
                         else
                         {
                             height = 125;
-                            width = (int)((tempFile.ImageWidth / tempFile.ImageHeight) * 125);
+                            width = (int)(((double)tempFile.ImageWidth / tempFile.ImageHeight) * 125);
                         }
 
                         await Common.RunStreamCommandAsync("ffmpeg",
